fix(app): harden anchored announcements
